requirements.txt
Telethon==1.34.0
google-generativeai==0.8.3
python-dotenv==1.0.0
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.6.15
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
rsa==4.9.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
